Step 4 (drift fix): move controls/budget/ → engine/budget/ and controls/pacing/ → engine/pacing/
Per DIR-PACKAGE-DRIFT-FIX-punit step 4. The controls/ catch-all
collapses; budget orchestration / monitor / formatter classes and
the pacing-configuration carrier all live next to the per-run state
holder engine/BudgetTracker, where they belong.

Touched:
- git mv 6 main + 6 test files plus package-info.java from
  controls/budget/ to engine/budget/.
- git mv 1 main file plus package-info.java from controls/pacing/
  to engine/pacing/.
- Rewrite package declarations on each moved file.
- Bulk-rewrite import paths across all four punit modules and
  punitexamples (~26 files).
- Drop the dead "org.javai.punit.controls.." entry from
  CoreArchitectureTest.utilPackageMustBeSelfContained's package list
  and the equivalent in ExampleArchitectureTest.
  PackageStructureArchitectureTest.noControlsPackage stays as a
  permanent regression guard now that its freeze-store has emptied.
- Drop the controls row from docs/DEVELOPER-API.md's
  package-visibility table.
- docs/DOMAIN-ONTOLOGY.md: refresh two package: lines
  (controls.budget → engine.budget; controls.pacing → engine.pacing).
- Empty controls/ tree (main and test) deleted.

Architecture-test bookkeeping:
- noControlsPackage rule's freeze-store: empty — at target state.
- engineUnderInternal rule's freeze-store: grew from 43 → 54 entries
  (the 7 budget + pacing main classes are current engine residents).
  Refrozen for this step (intermediate state for the cleanup arc;
  empties when the internal-namespace directive lands).
- Other rules' stores: unchanged.

Tests green across all four punit modules; punitexamples composite
build green.

Closes ArchUnit rule: noControlsPackage.
