+++ docs/DEVELOPER-API.md
@@ -440,7 +440,6 @@ The packages an author imports and what each one carries.
 | `org.javai.punit.engine.*` (criteria, baseline, explore, optimize, covariate, output) | Engine internals and concrete criterion impls                                                                                                 | **no** — internal                                             |
 | `org.javai.punit.contract.*`                                                          | Dead parallel stack, scheduled for deletion                                                                                                   | **no** — do not reference                                     |
 | `org.javai.punit.spec.*`                                                              | Internal spec model / registry / expiration / criteria                                                                                        | **no** — internal                                             |
-| `org.javai.punit.controls.*`                                                          | Internal budget / pacing implementations                                                                                                      | **no** — internal                                             |
 | `org.javai.punit.reporting`                                                           | Internal rendering helpers                                                                                                                    | **no** — internal                                             |
 | `org.javai.punit.util`                                                                | Internal utilities                                                                                                                            | **no** — internal                                             |
 | `org.javai.punit.junit5.*` (in punit-junit5)                                          | JUnit 5 extensions; one author-facing type (`UseCaseProvider`); annotations live in `punit-core`                                              | yes — for `UseCaseProvider` only                              |
