+++ punit-core/archunit_store/b705f7c8-ab73-4b68-8e35-130a0cf7f515
@@ -28,6 +28,16 @@ Class <org.javai.punit.engine.baseline.PowerAnalysis$BaselineLookup> does reside
 Class <org.javai.punit.engine.baseline.PowerAnalysis> does reside in a package 'org.javai.punit.engine..' in (PowerAnalysis.java:0)
 Class <org.javai.punit.engine.baseline.ProfileBoundBaselineProvider> does reside in a package 'org.javai.punit.engine..' in (ProfileBoundBaselineProvider.java:0)
 Class <org.javai.punit.engine.baseline.YamlBaselineProvider> does reside in a package 'org.javai.punit.engine..' in (YamlBaselineProvider.java:0)
+Class <org.javai.punit.engine.budget.BudgetMessageFormatter$1> does reside in a package 'org.javai.punit.engine..' in (BudgetMessageFormatter.java:0)
+Class <org.javai.punit.engine.budget.BudgetMessageFormatter> does reside in a package 'org.javai.punit.engine..' in (BudgetMessageFormatter.java:0)
+Class <org.javai.punit.engine.budget.BudgetOrchestrator$BudgetCheckResult> does reside in a package 'org.javai.punit.engine..' in (BudgetOrchestrator.java:0)
+Class <org.javai.punit.engine.budget.BudgetOrchestrator> does reside in a package 'org.javai.punit.engine..' in (BudgetOrchestrator.java:0)
+Class <org.javai.punit.engine.budget.CostBudgetMonitor$TokenMode> does reside in a package 'org.javai.punit.engine..' in (CostBudgetMonitor.java:0)
+Class <org.javai.punit.engine.budget.CostBudgetMonitor> does reside in a package 'org.javai.punit.engine..' in (CostBudgetMonitor.java:0)
+Class <org.javai.punit.engine.budget.DefaultTokenChargeRecorder> does reside in a package 'org.javai.punit.engine..' in (DefaultTokenChargeRecorder.java:0)
+Class <org.javai.punit.engine.budget.SharedBudgetMonitor$Scope> does reside in a package 'org.javai.punit.engine..' in (SharedBudgetMonitor.java:0)
+Class <org.javai.punit.engine.budget.SharedBudgetMonitor> does reside in a package 'org.javai.punit.engine..' in (SharedBudgetMonitor.java:0)
+Class <org.javai.punit.engine.budget.SuiteBudgetManager> does reside in a package 'org.javai.punit.engine..' in (SuiteBudgetManager.java:0)
 Class <org.javai.punit.engine.covariate.CovariateResolver$Builder> does reside in a package 'org.javai.punit.engine..' in (CovariateResolver.java:0)
 Class <org.javai.punit.engine.covariate.CovariateResolver> does reside in a package 'org.javai.punit.engine..' in (CovariateResolver.java:0)
 Class <org.javai.punit.engine.covariate.PeriodMatcher$Parsed> does reside in a package 'org.javai.punit.engine..' in (PeriodMatcher.java:0)
@@ -41,3 +51,4 @@ Class <org.javai.punit.engine.optimize.OptimizationsResolver> does reside in a p
 Class <org.javai.punit.engine.optimize.OptimizeOutputWriter> does reside in a package 'org.javai.punit.engine..' in (OptimizeOutputWriter.java:0)
 Class <org.javai.punit.engine.output.LatencySection> does reside in a package 'org.javai.punit.engine..' in (LatencySection.java:0)
 Class <org.javai.punit.engine.output.ResultProjections> does reside in a package 'org.javai.punit.engine..' in (ResultProjections.java:0)
+Class <org.javai.punit.engine.pacing.PacingConfiguration> does reside in a package 'org.javai.punit.engine..' in (PacingConfiguration.java:0)
