--- punit-core/archunit_store/cd07f984-c9bc-473f-8624-454ace84b950
@@ -1,11 +1,0 @@
-Class <org.javai.punit.controls.budget.BudgetMessageFormatter$1> does reside in a package 'org.javai.punit.controls..' in (BudgetMessageFormatter.java:0)
-Class <org.javai.punit.controls.budget.BudgetMessageFormatter> does reside in a package 'org.javai.punit.controls..' in (BudgetMessageFormatter.java:0)
-Class <org.javai.punit.controls.budget.BudgetOrchestrator$BudgetCheckResult> does reside in a package 'org.javai.punit.controls..' in (BudgetOrchestrator.java:0)
-Class <org.javai.punit.controls.budget.BudgetOrchestrator> does reside in a package 'org.javai.punit.controls..' in (BudgetOrchestrator.java:0)
-Class <org.javai.punit.controls.budget.CostBudgetMonitor$TokenMode> does reside in a package 'org.javai.punit.controls..' in (CostBudgetMonitor.java:0)
-Class <org.javai.punit.controls.budget.CostBudgetMonitor> does reside in a package 'org.javai.punit.controls..' in (CostBudgetMonitor.java:0)
-Class <org.javai.punit.controls.budget.DefaultTokenChargeRecorder> does reside in a package 'org.javai.punit.controls..' in (DefaultTokenChargeRecorder.java:0)
-Class <org.javai.punit.controls.budget.SharedBudgetMonitor$Scope> does reside in a package 'org.javai.punit.controls..' in (SharedBudgetMonitor.java:0)
-Class <org.javai.punit.controls.budget.SharedBudgetMonitor> does reside in a package 'org.javai.punit.controls..' in (SharedBudgetMonitor.java:0)
-Class <org.javai.punit.controls.budget.SuiteBudgetManager> does reside in a package 'org.javai.punit.controls..' in (SuiteBudgetManager.java:0)
-Class <org.javai.punit.controls.pacing.PacingConfiguration> does reside in a package 'org.javai.punit.controls..' in (PacingConfiguration.java:0)
