+++ punit-core/src/main/java/org/javai/punit/engine/budget/package-info.java
@@ -22,4 +22,4 @@
  * @see org.javai.punit.api.ProbabilisticTestBudget
  * @see org.javai.punit.api.TokenChargeRecorder
  */
-package org.javai.punit.controls.budget;
+package org.javai.punit.engine.budget;
