+++ punit-core/src/main/java/org/javai/punit/engine/pacing/package-info.java
@@ -22,4 +22,4 @@
  *
  * @see org.javai.punit.api.Pacing
  */
-package org.javai.punit.controls.pacing;
+package org.javai.punit.engine.pacing;
