+++ punit-core/src/main/java/org/javai/punit/verdict/ProbabilisticTestVerdictBuilder.java
@@ -14,9 +14,9 @@ import org.javai.punit.api.spec.FailureCount;
 import org.javai.punit.api.spec.InconclusiveReasons;
 import org.javai.punit.api.spec.Verdict;
 import org.javai.punit.api.UseCaseAttributes;
-import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
-import org.javai.punit.controls.budget.SharedBudgetMonitor;
-import org.javai.punit.controls.pacing.PacingConfiguration;
+import org.javai.punit.engine.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.engine.budget.SharedBudgetMonitor;
+import org.javai.punit.engine.pacing.PacingConfiguration;
 import org.javai.punit.verdict.ExpirationStatus;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.reporting.RateFormat;
