+++ punit-core/src/main/java/org/javai/punit/verdict/VerdictAdapter.java
@@ -9,7 +9,7 @@ import org.javai.punit.api.spec.EngineRunSummary;
 import org.javai.punit.api.spec.EvaluatedCriterion;
 import org.javai.punit.api.spec.ProbabilisticTestResult;
 import org.javai.punit.api.spec.Verdict;
-import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.engine.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.engine.output.LatencySection;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.verdict.ProbabilisticTestVerdictBuilder.LatencyInput;
