+++ punit-core/src/test/java/org/javai/punit/engine/budget/SharedBudgetMonitorTest.java
@@ -1,4 +1,4 @@
-package org.javai.punit.controls.budget;
+package org.javai.punit.engine.budget;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import java.util.Optional;
