+++ punit-core/src/test/java/org/javai/punit/verdict/ProbabilisticTestVerdictBuilderTest.java
@@ -9,11 +9,11 @@ import java.util.Optional;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.ThresholdOrigin;
 import org.javai.punit.api.spec.Verdict;
-import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
-import org.javai.punit.controls.budget.SharedBudgetMonitor;
-import org.javai.punit.controls.budget.SharedBudgetMonitor.Scope;
+import org.javai.punit.engine.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.engine.budget.SharedBudgetMonitor;
+import org.javai.punit.engine.budget.SharedBudgetMonitor.Scope;
 import org.javai.punit.api.BudgetExhaustedBehavior;
-import org.javai.punit.controls.pacing.PacingConfiguration;
+import org.javai.punit.engine.pacing.PacingConfiguration;
 import org.javai.punit.verdict.ExpirationPolicy;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.spec.model.ExecutionSpecification;
