+++ punit-core/src/test/java/org/javai/punit/verdict/VerdictAdapterFidelityTest.java
@@ -20,7 +20,7 @@ import org.javai.punit.api.spec.FailureCount;
 import org.javai.punit.api.spec.FailureExemplar;
 import org.javai.punit.api.spec.ProbabilisticTestResult;
 import org.javai.punit.api.spec.Verdict;
-import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.engine.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.verdict.TerminationReason;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
