+++ punit-report/src/test/java/org/javai/punit/report/HtmlReportWriterTest.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.spec.FailureCount;
 import org.javai.punit.api.spec.FailureExemplar;
-import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.engine.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.engine.output.LatencySection;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
