+++ punit-report/src/test/java/org/javai/punit/report/VerdictXmlReaderTest.java
@@ -14,7 +14,7 @@ import java.util.Optional;
 import javax.xml.stream.XMLStreamException;
 
 import org.javai.punit.api.TestIntent;
-import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.engine.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.verdict.ExpirationStatus;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
