+++ punit-report/src/test/java/org/javai/punit/report/VerdictXmlWriterTest.java
@@ -24,7 +24,7 @@ import java.util.LinkedHashMap;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.spec.FailureCount;
 import org.javai.punit.api.spec.FailureExemplar;
-import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.engine.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.verdict.ExpirationStatus;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
