+++ punit-sentinel/src/test/java/org/javai/punit/sentinel/SentinelExecutorTest.java
@@ -9,7 +9,7 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.javai.punit.api.TestIntent;
-import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
+import org.javai.punit.engine.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
 import org.javai.punit.verdict.ProbabilisticTestVerdict;
